fix : Handler 인터페이스를 사용하는 방식으로 변경
- 기존의 State 루프를 도는것이 아니라 인터페이스를 참조하여 게임 루프가 동작하도록 변경했습니다.

+++ TextRpg/Data/JobData.cs
@@ -8,7 +8,7 @@ namespace TextRpg
     {
         public string Name { get; set; }
         public int MaxHP { get; set; }
-        public int Attack { get; set; }
-        public int Defense { get; set; }
+        public float Attack { get; set; }
+        public float Defense { get; set; }
     }
 }
